feat(main): refresh subcommand
The intention is to run it as a cronjob every 10 minutes to rebuild the
cached builds. This wasn't needed previously as there weren't quite so
many builds when we were limitting ourselves to only scrape "staging"
build branches. Now that we are scraping _all_ branches it's about ~3k
builds per day.